+++ nnpz/python/nnpz/io/output_column_providers/UniformPhotometry.py
@@ -1,4 +1,4 @@
-#
+ 
 # Copyright (C) 2012-2022 Euclid Science Ground Segment
 #
 # This library is free software; you can redistribute it and/or modify it under the terms of
@@ -69,8 +69,9 @@ class UniformPhotometry(OutputHandler.OutputColumnProviderInterface):
     def generate_output(self, indexes: np.ndarray, neighbor_info: np.ndarray, output: np.ndarray):
         neighbor_indexes = neighbor_info['NEIGHBOR_INDEX']
         neighbor_weight = neighbor_info['NEIGHBOR_WEIGHTS']
+        neighbor_scaling = neighbor_info['NEIGHBOR_SCALING']
         # Neighbor photometry on the target color space
-        ref_target_colorspace = neighbor_info['NEIGHBOR_PHOTOMETRY']
+        ref_target_colorspace = neighbor_info['CORRECTED_NEIGHBOR_PHOTOMETRY']
         target_photo = self.__target_phot[indexes]
         with np.errstate(divide='ignore', invalid='ignore'):
             for (col, col_err), (obj, obs, _, _) in self.__filter_map.items():
@@ -79,7 +80,7 @@ class UniformPhotometry(OutputHandler.OutputColumnProviderInterface):
                 ref_restframe = self.__ref_phot.get_fluxes(obj)[neighbor_indexes]
                 # Compute the ratio between the rest frame and the target colorspace
                 obs_idx = self.__ref_filters_idx[obs]
-                ratio = (ref_restframe / ref_target_colorspace[:, :, obs_idx, 0]).value
+                ratio = (ref_restframe / ref_target_colorspace[:, :, obs_idx, 0]*neighbor_scaling).value
                 # Normalize using the weights
                 ratio *= neighbor_weight
                 ratio /= np.sum(neighbor_weight, axis=-1)[..., np.newaxis]
@@ -88,3 +89,6 @@ class UniformPhotometry(OutputHandler.OutputColumnProviderInterface):
                 # observed flux into the rest frame color space)
                 output[col] = target_obs
                 np.multiply(output[col], ratio, out=output[col])
+                # Mask elements with inf error (ie nan input fluxes in bruteforce)
+                output[col][np.isinf(target_photo.get_fluxes(obs,return_error=True)[:,1])]=np.nan
+                output[col_err][np.isinf(target_photo.get_fluxes(obs,return_error=True)[:,1])]=np.nan