+++ nnpz/tests/python/io/output_column_properties/test_UniformPhotometry.py
@@ -34,12 +34,16 @@ def test_uniform_photometry(reference_photometry, target_photometry, reference_m
 
     neighbor_info = np.zeros(1, dtype=[('NEIGHBOR_INDEX', int, 2),
                                        ('NEIGHBOR_WEIGHTS', np.float32, 2),
-                                       ('NEIGHBOR_PHOTOMETRY', np.float32, (2, 3, 2))])
+                                       ('NEIGHBOR_SCALING', np.float32, 2),
+                                       ('NEIGHBOR_PHOTOMETRY', np.float32, (2, 3, 2)),
+                                       ('CORRECTED_NEIGHBOR_PHOTOMETRY', np.float32, (2, 3, 2))])
     assert neighbor_info['NEIGHBOR_PHOTOMETRY'].shape == reference_matched_photometry.shape
 
     neighbor_info['NEIGHBOR_INDEX'][0] = [0, 1]
     neighbor_info['NEIGHBOR_WEIGHTS'] = 1.
+    neighbor_info['NEIGHBOR_SCALING'] = 1.
     neighbor_info['NEIGHBOR_PHOTOMETRY'][0] = reference_matched_photometry
+    neighbor_info['CORRECTED_NEIGHBOR_PHOTOMETRY'][0] = reference_matched_photometry
     mock_output_handler.write_output_for([0], neighbor_info)
 
     columns = mock_output_handler.get_data_for_provider(uniform)